Updated base and os layers

+++ censorman_v2/src/base/base_cmdline.c
@@ -9,7 +9,7 @@ CmdLine cmdline_parse(Arena *arena, int argc, char *args[])
     cmdline.command = STR(args[0]);
     cmdline.args = list_create(arena, sizeof(CmdLineArg));
 
-    for(int i = 1; i < argc; ++i)
+    for(s64 i = 1; i < argc; ++i)
     {
         String arg_str = STR(args[i]);
         arg_str = string_trim(arg_str);
@@ -61,7 +61,7 @@ CmdLine cmdline_parse(Arena *arena, int argc, char *args[])
 b32 cmdline_has_flag(CmdLine *cmdline, String id)
 {
     u32 hash = hash_string(id, 0);
-    for(u64 i = 0; i < cmdline->args.count; ++i)
+    for(s64 i = 0; i < cmdline->args.count; ++i)
     {
         CmdLineArg *arg = (CmdLineArg *)list_get((void*)&cmdline->args, i);
         if(hash == arg->hash && arg->is_flag)
@@ -74,7 +74,7 @@ b32 cmdline_has_flag(CmdLine *cmdline, String id)
 
 b32 cmdline_has_any_flags(CmdLine *cmdline, StringArray ids)
 {
-    for(u64 i = 0; i < ids.count; ++i)
+    for(s64 i = 0; i < ids.count; ++i)
     {
         if(cmdline_has_flag(cmdline, ids.items[i]))
             return true;
@@ -85,7 +85,7 @@ b32 cmdline_has_any_flags(CmdLine *cmdline, StringArray ids)
 String cmdline_get_value(CmdLine *cmdline, String id)
 {
     u32 hash = hash_string(id, 0);
-    for(u64 i = 0; i < cmdline->args.count; ++i)
+    for(s64 i = 0; i < cmdline->args.count; ++i)
     {
         CmdLineArg *arg = (CmdLineArg *)list_get((void*)&cmdline->args, i);
         if(hash == arg->hash && arg->is_flag)
@@ -100,7 +100,7 @@ String cmdline_get_value(CmdLine *cmdline, String id)
 u64 cmdline_get_unflagged_count(CmdLine *cmdline)
 {
     u64 count = 0;
-    for(u64 i = 0; i < cmdline->args.count; ++i)
+    for(s64 i = 0; i < cmdline->args.count; ++i)
     {
         CmdLineArg *arg = (CmdLineArg *)list_get((void*)&cmdline->args, i);
         if(!arg->is_flag)
@@ -112,7 +112,7 @@ u64 cmdline_get_unflagged_count(CmdLine *cmdline)
 String cmdline_get_unflagged(CmdLine *cmdline, u64 index)
 {
     u64 arg_index = 0;
-    for(u64 i = 0; i < cmdline->args.count; ++i)
+    for(s64 i = 0; i < cmdline->args.count; ++i)
     {
         CmdLineArg *arg = (CmdLineArg *)list_get((void*)&cmdline->args, i);
         if(!arg->is_flag)
@@ -133,7 +133,7 @@ void cmdline_print(CmdLine *cmdline)
     StringList sl = string_list_create(temp.arena);
     string_list_addf(&sl, "\nCommand: " STR_FMT, STR_ARG(cmdline->command));
 
-    for(u64 i = 0; i < cmdline->args.count; ++i)
+    for(s64 i = 0; i < cmdline->args.count; ++i)
     {
         CmdLineArg *arg = (CmdLineArg *)list_get(&cmdline->args, i);
         string_list_addf(&sl, "\n[Arg %d]: hash: %08X, is_flag: %s, value: " STR_FMT, i, arg->hash, STR_BOOL(arg->is_flag), STR_ARG(arg->value));
